requirements.txt
nose
sphinx
paho-mqtt
RPi.GPIO
pyyaml
adafruit-dht
adafruit-mcp3008
